Feature/framework specific (#1)
* Minor improvements

* Update version to 2.1.0

* Update global.json

* Update appveyor.yml

+++ Deploy.ps1
@@ -2,8 +2,11 @@ $scriptPath = Split-Path -Parent $PSCommandPath
 $deployPath = Join-Path -Path $scriptPath -ChildPath ".deploy"
 $packagesDeployPath = Join-Path -Path $deployPath -ChildPath "packages"
 
-Write-Host "Clearing deploy folder: $deployPath"
-Remove-Item $deployPath -Recurse -Force
+if ( Test-Path -Path $deployPath -PathType Container )
+{
+    Write-Host "Clearing deploy folder: $deployPath"
+    Remove-Item $deployPath -Recurse -Force
+}
 
 Write-Host "Testing PlatformSpecific MSTests"
 dotnet test test/PlatformSpecific.MSTest.Tests/PlatformSpecific.MSTest.Tests.csproj -c Release
